11-224

+++ ProjectForge/ControlMessage.h
@@ -5,6 +5,6 @@ class ControlMessage
 public:
 	int ID;
 	int toward;
-	ControlMessage(int ID,double toward);
+	ControlMessage(int ID,int toward);
 };
